Checking in alpha prototype.
This is code that runs "on my computer." It may not run anywhere else.
There are some hardcoded values (serial port for the Arduino, and so on)
that I cannot account for... meaning, I have no idea how much of this
code *must* be on my laptop.

This prototype may serve as enough inspiration to do some more work
along these lines, or clean it up enough to make it usable for others.