A working configuration.
This works on the Uno and on the older Arduino as well.

I don't think I've hard-coded anything... you have to hit some
configuration URLs from the server before things "just work", but
that is appropriate. It is not yet worked into the interface.

I've re-checked-in the tools, which have to be relative to this server.
The repository is getting large. We may end up nuking it and cleaning
things up later.